Make native button stick to same spot always
We are using some CSS trickery to make the canvas use the maximum viewport width or height, whichever is smaller, while maintaining aspect ratio.

We are also putting a native HTML button on top of the canvas.  Using even more CSS trickery, and some calculations, this button sticks to the same spot over the canvas, regardless of whether the browser is more portrait or more landscape.

This technique is so obscure that I'm not convinced we should ship with it.  But we would be able to replicate its functionality in JavaScript.