Simplify responsive button layout
It turns out we can have the button dimensions and positions relative to the width of the parent container, which is an even easier way to make them stick to the right places.  We can use media queries to get the font size to adjust.

This means we can remove all JS involved in button layout for now.

This change leaves the scene responsive, but caps the maximum width at 970px for convenience, and also centers it.