Use React for UI
Beginnings of a model in which we keep the app running with its own state machine, rendering to a couple canvases at full frame rate, using its own model classes... but with React used to draw the HTML-based UI on top of it, for headers, footers, and the body of the app.  This first attempt just tells React to re-render 10 times per second.

In this first commit, the first scene is rendered with both its original UI and a React-render duplicate UI on top.  The React UI doesn't yet shrink text below certain widths, though.