Set mode in the state using the URL only once
We were continually looking at the URL to determine what app mode we were in, which wouldn't have worked when integrated into studio.code.org.  Instead, we now check the URL once at the beginning and set appMode and appModeVariant in state at that time, when running in the standalone version of the app.