Final responsive work (hopefully)
This revits responsive layout to keep the UI looking more consistent as the screen scales up and down.

Most notably, text now has a baseline size in px for the current screen size, and different pieces of text simply use percentage variants of that baseline.  Line height is a consistent multiplier, 1.3, of font size regardless of screen size.

Buttons now scale their padding relative to the width as well.

As for screen size, we just have four screen sizes to consider right now:

  - at least 1440x810
  - at least 1280x720
  - at least 1024x576
  - and anything smaller in which we show at least 770x433

It would be possible to add support for some larger screen sizes again, allowing the tutorial to fill even more of big screens.

Our baseline font size is an arbitrary 18px at 930px (which is what we show when the screen is at least 1024x576), and we scale up or down from that.