Fix loading before training
We encountered an interesting situation on the beginning of the prediction screen on "words" levels, in which the screen froze before we showed the loading UI, and then we just saw a flash of it afterwards.

The hunch was that the training, for a significantly large number of fish (say 50 "yes" and 50 "no"), was causing the call to state.trainer.train() to immediately block the main thread, preventing React from rendering the loading UI before the training began.

This fix addresses this issue by delaying the start of training by half a second in the cases where we do intend to show the loading UI.  It appears to resolve the issue on a machine that previously exhibited the issue.